"some" additional changes & updates

+++ src/main/java/com/example/application/data/repository/CompanyRepository.java
@@ -1,9 +1,8 @@
 package com.example.application.data.repository;
 
 import com.example.application.data.entity.Company;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyRepository extends JpaRepository<Company, Integer> {
+public interface CompanyRepository extends JpaRepository<Company, Long> {
 
 }
