+++ src/main/java/com/example/application/data/repository/StatusRepository.java
@@ -1,9 +1,8 @@
 package com.example.application.data.repository;
 
 import com.example.application.data.entity.Status;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusRepository extends JpaRepository<Status, Integer> {
+public interface StatusRepository extends JpaRepository<Status, Long> {
 
 }
